Refactors to React Router for navigation
Migrates from custom JavaScript page transition logic to `react-router` for declarative application routing.

Extracts individual page content (Stopwatch, Settings, Info) into dedicated React components, improving modularity. Introduces a `Layout` component to manage common UI elements, such as navigation buttons, across different routes.

Removes the monolithic `App.tsx` and related manual page display toggles in CSS, streamlining page management and enhancing maintainability.
