+++ index.html
@@ -17,15 +17,12 @@
     <script src="assets/js/uuid.js"></script>
     <!-- <script src="assets/js/app.js"></script> -->
     <script src="assets/js/settingStorage.js"></script>
-    <script src="assets/js/pageTransition.js"></script>
     <!-- <script src="assets/js/teamStorage.js"></script> -->
     <script src="assets/js/stopwatch.js"></script>
     <script>
 
         toDisplayHtml(msToArrTime(prepDuration));
 
-        //init page
-		routeToPage('stopWatch')
 		printAllLaps()
     </script>
 </body>
