+++ package.json
@@ -11,7 +11,8 @@
     },
     "dependencies": {
         "react": "^19.2.0",
-        "react-dom": "^19.2.0"
+        "react-dom": "^19.2.0",
+        "react-router": "^7.9.5"
     },
     "devDependencies": {
         "@biomejs/biome": "2.3.3",
